+++ samples/src/gyroid_module.cpp
@@ -19,14 +19,14 @@ namespace {
 using namespace manifold;
 
 struct Gyroid {
-  float operator()(vec3 p) const {
-    p -= glm::pi<float>() / 4;
+  double operator()(vec3 p) const {
+    p -= glm::pi<double>() / 4;
     return cos(p.x) * sin(p.y) + cos(p.y) * sin(p.z) + cos(p.z) * sin(p.x);
   }
 };
 
-Manifold RhombicDodecahedron(float size) {
-  Manifold box = Manifold::Cube(size * glm::sqrt(2.0f) * vec3(1, 1, 2), true);
+Manifold RhombicDodecahedron(double size) {
+  Manifold box = Manifold::Cube(size * glm::sqrt(2.0) * vec3(1, 1, 2), true);
   Manifold result = box.Rotate(90, 45) ^ box.Rotate(90, 45, 90);
   return result ^ box.Rotate(0, 0, 45);
 }
@@ -42,9 +42,9 @@ namespace manifold {
  * demonstrates the use of a Signed Distance Function (SDF) to create smooth,
  * complex manifolds.
  */
-Manifold GyroidModule(float size, int n) {
-  auto gyroid = [&](float level) {
-    const float period = glm::two_pi<float>();
+Manifold GyroidModule(double size, int n) {
+  auto gyroid = [&](double level) {
+    const double period = glm::two_pi<double>();
     return Manifold(MeshGL::LevelSet(Gyroid(), {vec3(-period), vec3(period)},
                                      period / n, level))
         .Scale(vec3(size / period));
@@ -52,6 +52,6 @@ Manifold GyroidModule(float size, int n) {
 
   Manifold result = (RhombicDodecahedron(size) ^ gyroid(-0.4)) - gyroid(0.4);
 
-  return result.Rotate(-45, 0, 90).Translate({0, 0, size / glm::sqrt(2.0f)});
+  return result.Rotate(-45, 0, 90).Translate({0, 0, size / glm::sqrt(2.0)});
 }
 }  // namespace manifold
